feat: remove some of the commit message rules (#5)
## What does this PR do?

<!--
Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.

If we didn't discuss your PR in Github issues there's a high chance it
will not be merged.

The following links the related issue to the PR
(https://docs.github.com/en/free-pro-team@latest/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword)
-->

This PR removes some of the commit message rules. Now only it checks the
tag of the issue. e.g
Commit message format:  `tag: commit message`

tag can be : ["feat", "fix", "test", "chore"]

### Does your PR introduce any breaking changes? If yes, please list
them.

<!-- FILL IN or None -->

## Before submitting

- [ ] Was this **discussed/approved** via a GitHub issue? (not for typos
and docs)
- [ ] Did you make sure your **PR does only one thing**, instead of
bundling different changes together?
- [ ] Did you make sure to **update the documentation** with your
changes? (if necessary)
- [ ] Did you write any **new necessary tests**? (not for typos and
docs)
- [x] Did you verify new and **existing tests pass** locally with your
changes?
- [ ] Did you list all the **breaking changes** introduced by this pull
request?
- [ ] Did you **update the CHANGELOG**? (not for typos, docs, test
updates, or minor internal changes/refactors)

<!-- In the CHANGELOG, separate each item in the unreleased section by a
blank line to reduce collisions -->

Co-authored-by: pyler <psycoder@pyler.tech>
Co-authored-by: MLlead_steve <98260451+psycoder21@users.noreply.github.com>